Add Q16

+++ basic_conditions/q15.py
@@ -1,6 +1,6 @@
-a = float(input("Enter the number 1: "))
-b = float(input("Enter the number 2: "))
-c = float(input("Enter the number 3: "))
+a = int(input("Enter the number 1: "))
+b = int(input("Enter the number 2: "))
+c = int(input("Enter the number 3: "))
 
 if all(value > 0 for value in [a, b, c]):
     if a + b > c and a + c > b and b + c > a:
